Addresses #270 by adding retry logic to replica connections.
Under certain network conditions, it's possible for a connection from devpi to
replicas (such as pypi) to be dropped.

Before this commit, when this condition happens a FatalResponse exception is
generated resulting in a "502 bad gateway" error being returned by devpi to the
client. In many cases, this network error is transient and can be resolved by
retrying the request to the replica.

Currently, the requests library defaults to 0 retries for connection errors.

This commit updates devpi-common and devpi-server to enable a retry limit other
than 0. The retry limit is now specified in the configuration flag
"--replica-max-retries" and defaults to 5 retries.