Introduce stage customizer in preparation for plugins.
This is not doing anything yet.
The idea is that specific methods on the customizer class will be called to change the behaviour of the PrivateStage class.
This was done to avoid having plugins create it's whole own Stage class or subclass from an existing one, which would be hard to maintain.
The customizer class will have a defined interface.