Move import subscriber call outside of the import transaction.
When keyfs of the currently being imported transaction were accessed in the read transaction for the import subscriber, wrong values were used and stored in the cache.
Now the import subscriber can't prevent the import of the current serial anymore, but this isn't something we want/need anyway (see adjusted test_import_changes_subscriber_error).