Avoid accessing PyPI for projects that don't exist
Previously, when a project without any cached links was requested from
the `PyPIStage`, it would immediately try to load that package from the
upstream server. This change introduces a check to make sure the package
is actually in the stage before trying to load it.

This improves response times on slow connections, reduces the load on
the upstream server, and avoids leaking potential typos of private
package names to the upstream server.