Properly set PATH_INFO when X-outside-url header with path is used
When the `X-outside-url` header is passed in a request to devpi-server,
the path portion is set as the wsgi `SCRIPT_NAME` variable so it can
run in a subpath. However, the wsgi spec [1] states that `SCRIPT_NAME`
and `PATH_INFO` together constitute the complete URL but `PATH_INFO` isn't
modified. This commit adds stripping of the `SCRIPT_NAME` from the
beginning of `PATH_INFO`.

[1] https://wsgi.readthedocs.io/en/latest/definitions.html#standard-environ-keys