Fix using old credential cache for first time
When [async_]httpget() fails with a 401/403, it checks to see if it has
(or can generate) any new auth credentials to try. If so, it should
retry its request. There was a bug in the previous commit when moving
from locally cached credentials to storing them on the Xom, where the
check is on the old local cache. Thus, even when valid credentials
were generated, the first request would fail as the old cache was still
empty. Subsequent requests would then work correctly. Now we correctly
check the new cache.